requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
cryptography==37.0.2
customerio==1.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.50.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-environ==0.8.1
django-phone-field==1.8.1
django-ulid==0.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
filelock==3.8.0
identify==2.5.3
idna==3.3
install==1.3.5
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
nodeenv==1.7.0
oauthlib==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
ulid-py==1.1.0
urllib3==1.26.9
uWSGI==2.0.20
virtualenv==20.16.3
wcwidth==0.2.5
